compile buildsupport with -no-pie
Recent versions of gcc are compiled by default with -pie (position-independent executable) enabled, which makes the symbol addresses in binaries depending on an offset which is only known at runtime. The side effect is that addr2line becomes unable to locate the symbols to display traceback in case of a runtime exception.